refactor: modularize and relocate chat handler to chat-decoration feature
