+++ behaviour_pack/scripts-dev/events/index.ts
@@ -1,5 +1,5 @@
 import load_block_event_handler from "../features/interactions/block-interact";
-import load_chat_handler from "./chat";
+import load_chat_handler from "../features/chat-decoration/chat";
 import load_connections_handler from "./connections";
 import load_entity_event_handler from "../features/interactions/entity-die";
 import load_script_event_handler from "../features/interactions/script_events";
