+++ behaviour_pack/scripts-dev/features/chat-decoration/chat.ts
@@ -1,9 +1,7 @@
-import {EntityComponentTypes, EquipmentSlot, system, world} from "@minecraft/server";
-import api from "../api";
+import {system, world} from "@minecraft/server";
+import api from "../../api";
 
 export default function load_chat_handler() {
-
-    // Relay in-game chat and decorate chat
     world.beforeEvents.chatSend.subscribe((chat_event) => {
         const gamertag = chat_event.sender.name
         const thorny_user = api.ThornyUser.fetch_user(gamertag)
