+++ behaviour_pack/scripts-dev/features/chat-decoration/index.ts
@@ -1,0 +1,5 @@
+import load_chat_handler from "./chat";
+
+export default function loadChatDecorationFeature() {
+    load_chat_handler()
+}
